Affichage de plusieurs produit dans le panier

+++ app/Models/Produit_Commande.php
@@ -33,19 +33,29 @@ class Produit_Commande extends Model
     }
 
     static function ajoutProduit(\App\Models\Commande $commande, \App\Models\Produit $produit){
-        self::create([
-            "QUANTITE"=>1,
-            "TAILLE"=>0,
-            "ID_PRODUIT"=>$produit["ID"],
-            "ID_COMMANDE"=>$commande["ID"],
-            "PRIX"=>null
-        ]);
+        if(self::where(["ID_PRODUIT"=>$produit["ID"],"ID_COMMANDE"=>$commande["ID"]])->exists()){
+            $quantite = self::where(["ID_PRODUIT"=>$produit["ID"],"ID_COMMANDE"=>$commande["ID"]])->firstOrFail()["QUANTITE"];
+            self::where(["ID_PRODUIT"=>$produit["ID"],"ID_COMMANDE"=>$commande["ID"]])->update(["QUANTITE"=>$quantite+1]);
+        }else{
+            self::create([
+                "QUANTITE"=>1,
+                "TAILLE"=>0,
+                "ID_PRODUIT"=>$produit["ID"],
+                "ID_COMMANDE"=>$commande["ID"],
+                "PRIX"=>null
+            ]);
+        }
     }
 
     static function supprimerProduit(\App\Models\Commande $commande, \App\Models\Produit $produit){
         if($commande["ETAT"] != "panier"){
             throw \App\Models\Exceptions::createError(519);
         }
-        self::where(["ID_COMMANDE"=>$commande["ID"],"ID_PRODUIT"=>$produit["ID"]])->delete();
+        if(self::where(["ID_COMMANDE"=>$commande["ID"],"ID_PRODUIT"=>$produit["ID"]])->firstOrFail()["QUANTITE"] == 1){
+            self::where(["ID_COMMANDE"=>$commande["ID"],"ID_PRODUIT"=>$produit["ID"]])->delete();
+        }else{
+            $quantite = self::where(["ID_PRODUIT"=>$produit["ID"],"ID_COMMANDE"=>$commande["ID"]])->firstOrFail()["QUANTITE"];
+            self::where(["ID_PRODUIT"=>$produit["ID"],"ID_COMMANDE"=>$commande["ID"]])->update(["QUANTITE"=>$quantite-1]);
+        }
     }
 }
